Rules
The objective of Congkak is to move your shells into the storehouse that is situated on your left-hand side. You move your shells by picking up all of the shells in one of your houses and depositing one into each of the houses to the left of the space. The first person to empty all of his or her houses is the winner.